Move patient related classes to '/patient' package
