+++ app/src/main/java/org/resolvetosavelives/red/AppDatabase.kt
@@ -4,14 +4,14 @@ import android.arch.persistence.room.Database
 import android.arch.persistence.room.RoomDatabase
 import android.arch.persistence.room.TypeConverters
 import org.resolvetosavelives.red.bp.BloodPressureMeasurement
-import org.resolvetosavelives.red.search.Gender
-import org.resolvetosavelives.red.search.Patient
-import org.resolvetosavelives.red.search.PatientAddress
-import org.resolvetosavelives.red.search.PatientPhoneNumber
-import org.resolvetosavelives.red.search.PatientPhoneNumberType
-import org.resolvetosavelives.red.search.PatientStatus
-import org.resolvetosavelives.red.search.PatientWithAddressAndPhone
-import org.resolvetosavelives.red.search.SyncStatus
+import org.resolvetosavelives.red.patient.Gender
+import org.resolvetosavelives.red.patient.Patient
+import org.resolvetosavelives.red.patient.PatientAddress
+import org.resolvetosavelives.red.patient.PatientPhoneNumber
+import org.resolvetosavelives.red.patient.PatientPhoneNumberType
+import org.resolvetosavelives.red.patient.PatientStatus
+import org.resolvetosavelives.red.patient.PatientWithAddressAndPhone
+import org.resolvetosavelives.red.patient.SyncStatus
 import org.resolvetosavelives.red.util.InstantRoomTypeConverter
 import org.resolvetosavelives.red.util.LocalDateRoomTypeConverter
 import org.resolvetosavelives.red.util.UuidRoomTypeConverter
