+++ app/src/main/java/org/resolvetosavelives/red/newentry/phone/PatientPhoneEntryScreen.kt
@@ -13,7 +13,7 @@ import kotterknife.bindView
 import org.resolvetosavelives.red.R
 import org.resolvetosavelives.red.TheActivity
 import org.resolvetosavelives.red.router.screen.ScreenRouter
-import org.resolvetosavelives.red.search.OngoingPatientEntry
+import org.resolvetosavelives.red.patient.OngoingPatientEntry
 import org.resolvetosavelives.red.widgets.ScreenCreated
 import org.resolvetosavelives.red.widgets.setTextAndCursor
 import org.resolvetosavelives.red.widgets.showKeyboard
