+++ app/src/main/java/org/resolvetosavelives/red/patient/Age.kt
@@ -1,4 +1,4 @@
-package org.resolvetosavelives.red.search
+package org.resolvetosavelives.red.patient
 
 import org.threeten.bp.Instant
