+++ app/src/main/java/org/resolvetosavelives/red/sync/patient/PatientSync.kt
@@ -6,8 +6,8 @@ import io.reactivex.Observable
 import io.reactivex.Single
 import io.reactivex.rxkotlin.toCompletable
 import io.reactivex.schedulers.Schedulers.single
-import org.resolvetosavelives.red.search.PatientRepository
-import org.resolvetosavelives.red.search.SyncStatus
+import org.resolvetosavelives.red.patient.PatientRepository
+import org.resolvetosavelives.red.patient.SyncStatus
 import org.resolvetosavelives.red.sync.SyncConfig
 import org.resolvetosavelives.red.util.Just
 import org.resolvetosavelives.red.util.None
